Move BigQuery classes to GVS packages
