+++ src/main/java/org/broadinstitute/hellbender/utils/gvs/bigquery/StorageAPIAvroReaderAndBigQueryStatistics.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.bigquery;
+package org.broadinstitute.hellbender.utils.gvs.bigquery;
 
 import com.google.cloud.bigquery.JobStatistics;
