+++ src/main/java/org/broadinstitute/hellbender/utils/gvs/bigquery/TableReference.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.bigquery;
+package org.broadinstitute.hellbender.utils.gvs.bigquery;
 
 import com.google.common.collect.ImmutableList;
