+++ src/test/java/org/broadinstitute/hellbender/utils/bigquery/BigQueryUtilsUnitTest.java
@@ -3,6 +3,9 @@ package org.broadinstitute.hellbender.utils.bigquery;
 import com.google.cloud.bigquery.FieldValueList;
 import org.apache.avro.generic.GenericRecord;
 import org.broadinstitute.hellbender.GATKBaseTest;
+import org.broadinstitute.hellbender.utils.gvs.bigquery.BigQueryResultAndStatistics;
+import org.broadinstitute.hellbender.utils.gvs.bigquery.BigQueryUtils;
+import org.broadinstitute.hellbender.utils.gvs.bigquery.StorageAPIAvroReader;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import java.util.*;
