+++ src/test/java/org/broadinstitute/hellbender/utils/gvs/bigquery/PendingBQWriterTest.java
@@ -3,8 +3,6 @@ package org.broadinstitute.hellbender.utils.gvs.bigquery;
 import com.google.cloud.bigquery.FieldValueList;
 import com.google.protobuf.Descriptors;
 import org.broadinstitute.hellbender.GATKBaseTest;
-import org.broadinstitute.hellbender.utils.bigquery.BigQueryResultAndStatistics;
-import org.broadinstitute.hellbender.utils.bigquery.BigQueryUtils;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
